Split types.rs module
